Fix circular #include between util/stream/output.h and util/stream/debug.h commit_hash:4c198f344c22ca2387327f8ed6f9bed70d33f7e1
