+++ util/stream/output.h
@@ -263,6 +263,7 @@ static inline IOutputStream& operator<<(IOutputStream& o Y_LIFETIME_BOUND, wchar
 namespace NPrivate {
     IOutputStream& StdOutStream() noexcept;
     IOutputStream& StdErrStream() noexcept;
+    IOutputStream& StdDbgStream() noexcept;
 } // namespace NPrivate
 
 /**
@@ -279,6 +280,15 @@ namespace NPrivate {
  * Standard log stream.
  */
 #define Clog Cerr
+
+/**
+ * Standard debug stream.
+ *
+ * Behavior of this stream is controlled via `DBGOUT` environment variable.
+ * If this variable is set, then this stream is redirected into `stderr`,
+ * otherwise whatever is written into it is simply ignored.
+ */
+#define Cdbg (::NPrivate::StdDbgStream())
 
 /**
  * End-of-line output manipulator, basically the same as `std::endl`.
@@ -298,8 +308,32 @@ static inline void Flush(IOutputStream& o) {
  * Also see format.h for additional manipulators.
  */
 
-#include "debug.h"
-
 void RedirectStdioToAndroidLog(bool redirect);
 
-/** @} */
+/**
+ * Debug output stream. Writes into `stderr`.
+ */
+class TDebugOutput: public IOutputStream {
+public:
+    inline TDebugOutput() noexcept = default;
+    ~TDebugOutput() override = default;
+
+    TDebugOutput(TDebugOutput&&) noexcept = default;
+    TDebugOutput& operator=(TDebugOutput&&) noexcept = default;
+
+private:
+    void DoWrite(const void* buf, size_t len) override;
+};
+
+/**
+ * This function returns the current debug level as set via `DBGOUT` environment
+ * variable.
+ *
+ * Note that the proper way to use this function is via `Y_DBGTRACE` macro.
+ * There are very few cases when there is a need to use it directly.
+ *
+ * @returns                             Debug level.
+ * @see ETraceLevel
+ * @see DBGTRACE
+ */
+int StdDbgLevel() noexcept;
